refactor(e2e): group JS test app source by SDK major
Organize the test app around the SDK major it exercises instead of by
Next.js router. Client and function definitions move under src/v3/ and
src/v4/, and route files import from @/v3 and @/v4. Next still requires
route files under pages/ and app/, but the SDK boundary is now explicit
in the import paths rather than inferred from which router a file lives in.

Pure move + import-path rewrite; no behavior change.
