+++ tests/js/src/app/api/durable/async/route.ts
@@ -1,5 +1,5 @@
 import { step } from "inngest-v4";
-import { inngestV4 } from "@/inngest/client_v4";
+import { inngestV4 } from "@/v4/client";
 
 export const POST = inngestV4.endpoint(async () => {
   await step.sleep("brief-pause", "1s");
