+++ tests/js/src/v3/functions/sdk_cancel_test.ts
@@ -1,4 +1,4 @@
-import { inngest } from "@/inngest/client";
+import { inngest } from "@/v3/client";
 
 export const testCancel = inngest.createFunction(
   {
